consolidate use of http client; move deployment service response logic to controller
