+++ src/services/azuredevops.service.ts
@@ -1,11 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import axios from 'axios';
-import { IDeployment } from '../models/deployment.interface';
+import { HttpService } from '@nestjs/axios';
+import { firstValueFrom } from 'rxjs';
 
 @Injectable()
 export class AzureDevOpsService {
   private readonly platformUrl = 'https://vsrm.dev.azure.com';
   private readonly token = process.env.DORA_AZURE_TOKEN;
+
+  constructor(private readonly httpService: HttpService) {}
 
   private getHeaders() {
     return {
@@ -14,50 +16,19 @@ export class AzureDevOpsService {
     };
   }
 
-  async getDeployments(
-    org: string,
-    project: string,
-    pipeline: string,
-    env: string,
-  ) {
-    let deployments: IDeployment[] = [];
+  async getDeployments(org: string, project: string) {
     const encodedOrg = encodeURIComponent(org);
     const encodedProject = encodeURIComponent(project);
     const url = `${this.platformUrl}/${encodedOrg}/${encodedProject}/_apis/release/releases?api-version=7.1&$expand=environments`;
 
     try {
-      const response = await axios.get(url, { headers: this.getHeaders() });
-      const releases = response.data.value;
-
-      // Filter for specific pipeline
-      let filteredReleases = releases.filter(
-        (release) => release.releaseDefinition.name === pipeline,
+      const response = await firstValueFrom(
+        this.httpService.get(url, {
+          headers: this.getHeaders(),
+        }),
       );
+      const deployments = response.data.value;
 
-      // filter for specific environment
-      for (let i = 0; i < filteredReleases.length; ++i) {
-        let release = filteredReleases[i];
-
-        // find env of the release and make sure it has succeeded (filter out pending)
-        const releaseEnv = release.environments.find(
-          (environment) =>
-            environment.name.toLowerCase() === env.toLowerCase() &&
-            environment.status === 'succeeded',
-        );
-
-        if (releaseEnv) {
-          // deployment time is determined by timestamp of final step in deployment
-          if (releaseEnv.deploySteps && releaseEnv.deploySteps.length > 0) {
-            const lastStep =
-              releaseEnv.deploySteps[releaseEnv.deploySteps.length - 1];
-            if (lastStep && lastStep.lastModifiedOn) {
-              deployments.push({ deploymentDate: lastStep.lastModifiedOn });
-            }
-          }
-        }
-      }
-
-      // return deploymentsWithCompletionTime;
       return deployments;
     } catch (error) {
       console.error('Error fetching deployments:', error);
